+++ src/Tests-internal/ECS/Test_StructHeap.cs
@@ -13,8 +13,11 @@ public static class Test_StructHeap
         var schema = EntityStore.GetEntitySchema();
         
         var type    = schema.GetComponentType<MyComponent1>();
-        var heap    = (StructHeap<MyComponent1>)type.CreateHeap();
+        var heap    = (StructHeap<MyComponent1>)type.CreateHeap(ArchetypeUtils.MinCapacity);
         AreEqual(512, heap.components.Length);
+
+        var low     = (StructHeap<MyComponent1>)type.CreateHeap(64);
+        AreEqual(64, low.components.Length);
     }
 }